+++ forui/lib/widgets/date_field.dart
@@ -6,5 +6,7 @@
 library forui.widgets.date_field;
 
 export '../src/widgets/date_field/date_field.dart';
+export '../src/widgets/date_field/date_field_controller.dart'
+    hide InternalFDateFieldControl, InternalFDateFieldController, Lifted, Managed;
 export '../src/widgets/date_field/date_field_style.dart';
 export '../src/widgets/date_field/calendar/properties.dart';